refactor: Order domain refactor replace double with BigDecimal

+++ src/main/java/com/paymentteamproject/domain/orderProduct/dto/getOneOrderProductResponse.java
@@ -3,21 +3,22 @@ package com.paymentteamproject.domain.orderProduct.dto;
 import com.paymentteamproject.domain.order.consts.OrderStatus;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
 public class getOneOrderProductResponse {
     private final Long orderNumber;
     private final Long orderId;
-    private final double totalAmount;
-    private final double userPoints;
-    private final double finalAmount;
-    private final double earnedPoints;
+    private final BigDecimal totalAmount;
+    private final BigDecimal userPoints;
+    private final BigDecimal finalAmount;
+    private final BigDecimal earnedPoints;
     private final String currency;
     private final OrderStatus status;
     private final LocalDateTime createdAt;
 
-    public getOneOrderProductResponse(Long orderNumber, Long orderId, double totalAmount, double userPoints, double finalAmount, double earnedPoints, String currency, OrderStatus status, LocalDateTime createdAt) {
+    public getOneOrderProductResponse(Long orderNumber, Long orderId, BigDecimal totalAmount, BigDecimal userPoints, BigDecimal finalAmount, BigDecimal earnedPoints, String currency, OrderStatus status, LocalDateTime createdAt) {
         this.orderNumber = orderNumber;
         this.orderId = orderId;
         this.totalAmount = totalAmount;
